[feat] modify domain for deploy

See merge request s11-webmobile1-sub2/S11P12A806!242

+++ backend/src/main/java/com/apink/poppin/api/user/service/UserServiceImpl.java
@@ -96,7 +96,7 @@ public class UserServiceImpl implements UserService {
 
             userRepository.save(findUser);
 
-            if (oldImgPath != null && !oldImgPath.equals("/uploads/profile.png")) {
+            if (image != null && oldImgPath != null && !oldImgPath.equals("/uploads/profile.png")) {
                 fileStorageService.deleteFile(oldImgPath);
             }
